refactor(orchestrator): extract GraphRunner for LangGraph execution drivers
Move the workflow/review/planning execution drivers and their LangGraph
setup helpers out of OrchestratorService into a new GraphRunner collaborator
(amelia/server/orchestrator/runner.py). The service constructs one GraphRunner
via constructor DI and delegates each driver call through self._runner, while
retaining ownership of _active_tasks registration and cleanup callbacks.

Public driver methods: run_workflow, run_workflow_with_retry,
run_review_workflow, run_planning_task. Setup helpers (_create_server_graph,
_build_runnable_config, _create_sandbox_provider, _reconstruct_initial_state,
_resolve_prompts, _sync_plan_from_checkpoint, _get_profile_or_fail,
_update_profile_repo_root) move with them and are reused by lifecycle methods
that stay on the service via self._runner.

Relocate the checkpoint-resume, retry/backoff, and plan-sync tests to a new
test_runner.py targeting GraphRunner directly; retarget all patch/mock sites
across the unit and integration suites.
