+++ tests/integration/test_github_issue_workflow.py
@@ -131,7 +131,7 @@ def mock_graph(langgraph_mock_factory: Any):
     """Patch create_implementation_graph to return a mock graph."""
     mocks = langgraph_mock_factory(astream_items=[])
     with patch(
-        "amelia.server.orchestrator.service.create_implementation_graph"
+        "amelia.server.orchestrator.runner.create_implementation_graph"
     ) as mock_create_graph:
         mock_create_graph.return_value = mocks.graph
         yield mocks
