+++ tests/integration/test_queue_workflow_flow.py
@@ -56,7 +56,7 @@ def patched_graph(langgraph_mock_factory: Any):
     """
     mocks = langgraph_mock_factory(astream_items=[])
     with patch(
-        "amelia.server.orchestrator.service.create_implementation_graph"
+        "amelia.server.orchestrator.runner.create_implementation_graph"
     ) as mock_create_graph:
         mock_create_graph.return_value = mocks.graph
         yield mocks
